% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_task.R
\name{add_task}
\alias{add_task}
\title{add_task function}
\usage{
add_task(df = NULL)
}
\arguments{
\item{df:}{can be NULL, represents a data frame the user wants to add a task to}
}
\value{
returns a data frame
}
\description{
Creates a data frame for a user to then input tasks. If data frame supplied in parameters, then allows the user to add tasks to that data frame
}
\examples{
add_task()
add_task(to_do_list)
}
