% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_tasks.R
\name{add_tasks}
\alias{add_tasks}
\title{add_task function}
\usage{
add_tasks(df = NULL)
}
\arguments{
\item{df}{can be NULL, represents a data frame the user wants to add a task to}
}
\value{
returns a data frame
}
\description{
Creates a data frame for a user to then input tasks. If data frame supplied in parameters, then allows the user to add tasks to that data frame
}
\examples{
if (interactive()) {
  add_task()
}

}
