% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_task.R
\name{delete_task}
\alias{delete_task}
\title{Delete specific tasks from data frame}
\usage{
delete_task(tasks, description_name = NULL)
}
\arguments{
\item{tasks}{the name of the data frame}
}
\value{
An updated_tasks data frame that does not include the deleted tasks
}
\description{
Delete specific tasks from data frame
}
\examples{
delete_task(tasks)
}
