% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_tasks.R
\name{edit_tasks}
\alias{edit_tasks}
\title{The Edit function}
\usage{
edit_tasks(old_data_frame, row_number)
}
\arguments{
\item{dataframe, row_number, }{These inputs are all required so
that the function can edit the correct row.}
}
\value{
Outputs a readline where the user gets to input new values if needed
in their desired data frame. There is an option to "keep" the old values if
needed by typing in "keep", and if the user tries to input a value that does
not align with each column's params, a warning is issued.
}
\description{
The user inputs their desired data frame they want to edit, and the function
finds it by taking in the row number and returning options to edit the whole
row.
}
\examples{
tasks <- data.frame(
  Description = c("Task 1", "Task 2", "Task 3"),
  Due_Date = c("04-21-2024", "05-10-2025", "05-10-2025"),
  Priority = c("Medium", "High", "Low"),
  Category = c("SDS", "PSY", "CSC"),
  Completed = c(FALSE, TRUE, FALSE)
)

if (interactive()) {
  edit_task(tasks, 1)
}

}
