% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_task.R
\name{find_task}
\alias{find_task}
\title{The Find function}
\usage{
find_task(dataframe, phrase)
}
\arguments{
\item{dataframe, phrase}{Where "dataframe" is the data frame of interest and
"phrase" is the phrase (character) being searched for under the subset of
dataframe$Description}
}
\value{
A string that explains where the phrase is found and what the full
phrase is.
}
\description{
The user inputs their desired phrase they want to search for and find and
under which data frame it may be under, and the function returns under what
row it is. If there is multiple matches, it returns each one.
}
\examples{
tasks <- data.frame(
 Description = c("physics hw 1", "physics hw 2", "math 1"),
 Due_Date = as.Date(as.character("2024-04-18", "2024-04-19", "2024-04-19")),
 Priority = c("High", "Medium", "Low"),
 Category = c("Homework", "Homework", "Assignment"),
 Completed = c(FALSE, FALSE, TRUE)
)
tasks <- find_task(tasks, "hw")
}
