% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organize_function.R
\name{organize}
\alias{organize}
\title{Organize data frame by columns}
\usage{
organize(tasks)
}
\arguments{
\item{tasks:}{a data frame representing a list of tasks to do, usually made with add_task function}
}
\value{
returns a data frame
}
\description{
Given a data frame, which represents a to do list, organize the
to do tasks by columns named due date, completion status, or category
}
\examples{
organize(to_do_list)

          Description   Due_Date Priority Category Completed
1 finish final project 2024-05-10   High   School     FALSE
}
