% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organize_tasks.R
\name{organize_tasks}
\alias{organize_tasks}
\title{Organize data frame by columns}
\usage{
organize_tasks(tasks, choice = NULL)
}
\arguments{
\item{tasks}{a data frame representing a list of tasks to do, usually made
with add_task function}

\item{choice}{can be NULL, or 1,2,3, or 4. Represents what the user would like to choose to organize by}
}
\value{
returns a data frame
}
\description{
Given a data frame, which represents a to do list, organize the
to do tasks by columns named due date, completion status, or category
}
\examples{
organize(to_do_list)

          Description   Due_Date Priority Category Completed
1 finish final project 2024-05-10   High   School     FALSE
}
