% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\docType{data}
\name{sample_data}
\alias{sample_data}
\title{Sample Tasks Data}
\format{
\subsection{'sample_data'}{

A data frame with 8 rows and 5 columns:
\describe{
\item{Description}{Description of the task}
\item{Due_Date}{When the task is due (Month-Day-Year)}
\item{Priority}{The priority level of the task (High, Medium, Low)}
\item{Category}{The category the task falls under. It could be for a certain class or personal}
\item{Completed}{Whether the task is complete or incomplete}
}
}
}
\usage{
sample_data
}
\description{
This is a example of a data frame consisting of various tasks.
}
\keyword{datasets}
