% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_tasks.R
\name{split_function}
\alias{split_function}
\title{The split function}
\usage{
split_function(dataframe, column_name, cat_name)
}
\arguments{
\item{dataframe, column_name, cat_name}{the data frame being split up,
and under what column name and specific value(s) that are wanted to be
rounded up into a new data frame.}
}
\value{
This function returns a new data frame under the desired params.
}
\description{
A function where the user inputs what data frame they want to split and by
what param they want to split it by.
}
\examples{

tasks <- data.frame(
 Description = c("physics hw 1", "physics hw 2", "math 1"),
 Due_Date = as.Date(as.character("2024-04-18", "2024-04-19", "2024-04-19")),
 Priority = c("High", "Medium", "Low"),
 Category = c("Homework", "Homework", "Assignment"),
 Completed = c(FALSE, FALSE, TRUE)
)
split_function(tasks, "Category", "Homework")

}
