% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_tasks.R
\name{split_tasks}
\alias{split_tasks}
\title{The split_tasks function}
\usage{
split_tasks(dataframe, column_name, value_name, new_df_name = NULL)
}
\arguments{
\item{dataframe, column_name, value_name, new_df_name}{the data frame being split up,
and under what column name and specific value(s) that are wanted to be
rounded up into a new data frame, the name of the new data frame.}
}
\value{
This function returns a new data frame under the desired params.
}
\description{
A function where the user inputs what data frame they want to split and by
what param they want to split it by.
}
\examples{
tasks <- data.frame(
Description = c("Task 1", "Task 2", "Task 3"),
Due_Date = c("04-21-2024", "05-10-2025", "05-10-2025"),
Priority = c("Medium", "High", "Low"),
Category = c("SDS", "PSY", "CSC"),
Completed = c(FALSE, TRUE, FALSE)
)

if (interactive()) {
  split_tasks(tasks, "Priority", "High")
}

}
