% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_task.R
\name{delete_task}
\alias{delete_task}
\title{Delete specific tasks from data frame}
\usage{
delete_task(tasks, description_name = NULL)
}
\arguments{
\item{tasks, description_name}{the name of the data frame and the optional description of the task to be deleted}
}
\value{
An updated_tasks data frame that does not include the deleted tasks
}
\description{
Delete specific tasks from data frame
}
\examples{
tasks <- data.frame(
Description = c("Task 1", "Task 2", "Task 3"),
Due_Date = c("04-21-2024", "05-10-2025", "05-10-2025"),
Priority = c("Medium", "High", "Low"),
Category = c("SDS", "PSY", "CSC"),
Completed = c(FALSE, TRUE, FALSE)
)

if (interactive()) {
  delete_tasks(tasks)
}

}
