% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_task.R
\name{delete_task}
\alias{delete_task}
\title{Delete specific tasks from data frame}
\usage{
delete_task(tasks)
}
\arguments{
\item{tasks}{the name of the data frame}
}
\value{
An updated tasks data frame that does not include the deleted tasks
}
\description{
Delete specific tasks from data frame
}
\examples{

tasks <- data.frame(
 Description = c("physics hw 1", "physics hw 2", "math 1"),
 Due_Date = as.Date(as.character("2024-04-18", "2024-04-19", "2024-04-19")),
 Priority = c("High", "Medium", "Low"),
 Category = c("Homework", "Homework", "Assignment"),
 Completed = c(FALSE, FALSE, TRUE)
)
delete_task(tasks)
}
