% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary_tasks}
\alias{summary_tasks}
\title{Summarize and visualize key information from tasks}
\usage{
summary_tasks(tasks)
}
\arguments{
\item{tasks}{a dataframe}
}
\value{
This function does not return any value, it only prints information
}
\description{
Summarize and visualize key information from tasks
}
\examples{
summary_tasks(df)
}
